+++ locales/ar/common.json
@@ -234,7 +234,7 @@
   "exportType.allAgentWithMessage": "تصدير جميع الوكلاء والرسائل",
   "exportType.globalSetting": "تصدير الإعدادات العامة",
   "feedback": "ملاحظات",
-  "feedback.emailContact": "يمكنك أيضًا مراسلتنا عبر البريد الإلكتروني على {{email}}",
+  "feedback.emailContact": "يمكنك أيضًا مراسلتنا عبر البريد الإلكتروني على <email>{{email}}</email>",
   "feedback.errors.fileTooLarge": "الملف يتجاوز الحجم المسموح به (5 ميغابايت)",
   "feedback.errors.submitFailed": "فشل في الإرسال. حاول مرة أخرى.",
   "feedback.errors.teamNotFound": "خطأ في التكوين",
