+++ locales/en-US/common.json
@@ -234,7 +234,7 @@
   "exportType.allAgentWithMessage": "Export All Agents and Messages",
   "exportType.globalSetting": "Export Global Settings",
   "feedback": "Feedback",
-  "feedback.emailContact": "You can also email us at {{email}}",
+  "feedback.emailContact": "You can also email us at <email>{{email}}</email>",
   "feedback.errors.fileTooLarge": "File exceeds 5MB",
   "feedback.errors.submitFailed": "Submit failed. Try again.",
   "feedback.errors.teamNotFound": "Configuration error",
