+++ locales/it-IT/common.json
@@ -234,7 +234,7 @@
   "exportType.allAgentWithMessage": "Esporta tutti gli agenti e i messaggi",
   "exportType.globalSetting": "Esporta impostazioni globali",
   "feedback": "Feedback",
-  "feedback.emailContact": "Puoi anche inviarci un'email a {{email}}",
+  "feedback.emailContact": "Puoi anche inviarci un'email a <email>{{email}}</email>",
   "feedback.errors.fileTooLarge": "Il file supera i 5MB",
   "feedback.errors.submitFailed": "Invio non riuscito. Riprova.",
   "feedback.errors.teamNotFound": "Errore di configurazione",
