+++ locales/ja-JP/common.json
@@ -234,7 +234,7 @@
   "exportType.allAgentWithMessage": "すべてのアシスタントとメッセージをエクスポート",
   "exportType.globalSetting": "グローバル設定をエクスポート",
   "feedback": "フィードバック",
-  "feedback.emailContact": "また、{{email}} にメールを送ることもできます",
+  "feedback.emailContact": "また、<email>{{email}}</email> にメールを送ることもできます",
   "feedback.errors.fileTooLarge": "ファイルサイズが5MBを超えています",
   "feedback.errors.submitFailed": "送信に失敗しました。もう一度お試しください。",
   "feedback.errors.teamNotFound": "設定エラー",
