+++ locales/pl-PL/common.json
@@ -234,7 +234,7 @@
   "exportType.allAgentWithMessage": "Eksportuj wszystkich agentów i wiadomości",
   "exportType.globalSetting": "Eksportuj ustawienia globalne",
   "feedback": "Opinie",
-  "feedback.emailContact": "Możesz również wysłać do nas e-mail na adres {{email}}",
+  "feedback.emailContact": "Możesz również wysłać do nas e-mail na adres <email>{{email}}</email>",
   "feedback.errors.fileTooLarge": "Plik przekracza 5 MB",
   "feedback.errors.submitFailed": "Nie udało się wysłać. Spróbuj ponownie.",
   "feedback.errors.teamNotFound": "Błąd konfiguracji",
