+++ locales/ru-RU/common.json
@@ -234,7 +234,7 @@
   "exportType.allAgentWithMessage": "Экспорт всех агентов и сообщений",
   "exportType.globalSetting": "Экспорт глобальных настроек",
   "feedback": "Обратная связь",
-  "feedback.emailContact": "Вы также можете написать нам на {{email}}",
+  "feedback.emailContact": "Вы также можете написать нам на <email>{{email}}</email>",
   "feedback.errors.fileTooLarge": "Файл превышает 5 МБ",
   "feedback.errors.submitFailed": "Не удалось отправить. Попробуйте ещё раз.",
   "feedback.errors.teamNotFound": "Ошибка конфигурации",
