+++ locales/zh-CN/common.json
@@ -234,7 +234,7 @@
   "exportType.allAgentWithMessage": "导出所有助理和消息",
   "exportType.globalSetting": "导出全局设置",
   "feedback": "反馈与建议",
-  "feedback.emailContact": "您也可以发送邮件至 {{email}}",
+  "feedback.emailContact": "您也可以发送邮件至 <email>{{email}}</email>",
   "feedback.errors.fileTooLarge": "文件大小超过 5MB 限制",
   "feedback.errors.submitFailed": "提交失败，请重试",
   "feedback.errors.teamNotFound": "配置错误",
