+++ src/components/FeedbackModal/types.ts
@@ -1,0 +1,4 @@
+export interface FeedbackInitialValues {
+  message?: string;
+  title?: string;
+}
