+++ src/features/CommandMenu/MainMenu.tsx
@@ -13,9 +13,9 @@ import {
 import { memo } from 'react';
 import { useTranslation } from 'react-i18next';
 
+import { openFeedbackModal } from '@/components/FeedbackModal';
 import { getNavigableRoutes, getRouteById } from '@/config/routes';
 import { FEEDBACK } from '@/const/url';
-import { useFeedbackModal } from '@/hooks/useFeedbackModal';
 
 import { useCommandMenuContext } from './CommandMenuContext';
 import { CommandItem } from './components';
@@ -25,7 +25,6 @@ import { useCommandMenu } from './useCommandMenu';
 const MainMenu = memo(() => {
   const { pathname, menuContext, setPages, pages } = useCommandMenuContext();
   const { t } = useTranslation('common');
-  const { open: openFeedbackModal } = useFeedbackModal();
 
   const {
     handleCreateSession,
