+++ src/routes/(main)/_layout/index.tsx
@@ -5,7 +5,7 @@ import { TITLE_BAR_HEIGHT } from '@lobechat/desktop-bridge';
 import { Flexbox } from '@lobehub/ui';
 import { cx } from 'antd-style';
 import { type FC } from 'react';
-import { lazy, Suspense } from 'react';
+import { Suspense } from 'react';
 import { HotkeysProvider } from 'react-hotkeys-hook';
 import { Outlet } from 'react-router-dom';
 
@@ -24,7 +24,6 @@ import TitleBar from '@/features/Electron/titlebar/TitleBar';
 import HotkeyHelperPanel from '@/features/HotkeyHelperPanel';
 import NavPanel from '@/features/NavPanel';
 import { RouteMetaBridge } from '@/features/RouteMeta';
-import { useFeedbackModal } from '@/hooks/useFeedbackModal';
 import { usePlatform } from '@/hooks/usePlatform';
 import { MarketAuthProvider } from '@/layout/AuthProvider/MarketAuth';
 import CmdkLazy from '@/layout/GlobalProvider/CmdkLazy';
@@ -39,18 +38,11 @@ import DesktopLayoutContainer from './DesktopLayoutContainer';
 import RegisterHotkeys from './RegisterHotkeys';
 import { styles } from './style';
 
-const FeedbackModal = lazy(() => import('@/components/FeedbackModal'));
-
 const CloudBanner = dynamic(() => import('@/features/AlertBanner/CloudBanner'));
 
 const Layout: FC = () => {
   const { isPWA } = usePlatform();
   const { showCloudPromotion } = useServerConfigStore(featureFlagsSelectors);
-  const {
-    initialValues: feedbackInitialValues,
-    isOpen: isFeedbackModalOpen,
-    close: closeFeedbackModal,
-  } = useFeedbackModal();
 
   return (
     <HotkeysProvider initiallyActiveScopes={[HotkeyScopeEnum.Global]}>
@@ -99,15 +91,6 @@ const Layout: FC = () => {
         <HotkeyHelperPanel />
         <RegisterHotkeys />
         <CmdkLazy />
-        {isFeedbackModalOpen && (
-          <Suspense fallback={null}>
-            <FeedbackModal
-              initialValues={feedbackInitialValues}
-              open={isFeedbackModalOpen}
-              onClose={closeFeedbackModal}
-            />
-          </Suspense>
-        )}
       </Suspense>
     </HotkeysProvider>
   );
