+++ src/routes/(mobile)/me/(home)/features/useCategory.tsx
@@ -15,13 +15,14 @@ import { useNavigate } from 'react-router-dom';
 
 import useBusinessMeCells from '@/business/client/features/User/useBusinessMeCells';
 import { type CellProps } from '@/components/Cell';
+import { openChangelogModal } from '@/components/ChangelogModal';
 import { DOCUMENTS, FEEDBACK } from '@/const/index';
 import { usePlatform } from '@/hooks/usePlatform';
 import { featureFlagsSelectors, useServerConfigStore } from '@/store/serverConfig';
 import { useUserStore } from '@/store/user';
 import { authSelectors } from '@/store/user/selectors';
 
-export const useCategory = (onOpenChangelogModal: () => void) => {
+export const useCategory = () => {
   const navigate = useNavigate();
   const { t } = useTranslation(['common', 'setting', 'auth']);
   const { showCloudPromotion, hideDocs } = useServerConfigStore(featureFlagsSelectors);
@@ -91,7 +92,7 @@ export const useCategory = (onOpenChangelogModal: () => void) => {
       icon: FileClockIcon,
       key: 'changelog',
       label: t('changelog'),
-      onClick: onOpenChangelogModal,
+      onClick: () => openChangelogModal(),
     },
   ].filter(Boolean) as CellProps[];
